requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
daphne==3.0.2
distlib==0.3.4
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.6.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.4
virtualenvwrapper-win==1.2.7
zope.interface==5.4.0
